requirements.txt
langchain==0.1.13
python-dotenv==0.19.1
pydantic~=1.10.14
langchain-openai
